9043/refactor: move PaymentCalculatorService to the other folder
